Move "LocallyGenerated..." messages
